+++ CHANGELOG.md
@@ -118,3 +118,5 @@ was removed or the new one was added to folder
 - Colors can be initialized with enum
 - CreateTempFont from TTF repaired and should be used as de facto standard instead of create font
 - HasTag function for TagUiSection
+- File explorer and Path Maker also use ttf fonts
+- TTFFont creation supports LocalTexMan